requirements.txt
aniso8601==4.1.0
apipkg==1.5
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
execnet==1.5.0
Flask==1.0.2
Flask-JWT-Extended==3.17.0
Flask-RESTful==0.3.7
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
nose==1.3.7
pluggy==0.8.1
psycopg2-binary==2.7.7
py==1.7.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.2.2
pytest==4.2.0
pytest-cov==2.6.1
pytest-forked==1.0.1
pytest-xdist==1.26.1
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
